==> 1_Original_Tweet_Data_Acquistion_TwitterAPI.ipynb <==
Author: Hnin Pwint Tin, 20 May 2021

Social and Information Network Analysis, 2021 Autumn

Master of Data Science and Innovation , University of Technology Sydney

This is the data preparation script part-1 for network data visualisation assessment task - AT3

import os
import tweepy as tw
import pandas as pd

import csv
import datetime

### Twitter Developer Account Credentials
consumer_key = 'xxxx'
consumer_secret = "xxx"
access_token = "xx-xx"
access_token_secret = "xxxx"

auth = tw.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_token_secret)

api = tw.API(auth, wait_on_rate_limit=True)

search_words = "covidvaccination"

filepath = "data/Hnin_Tweeter_Api/"
db_tweets = pd.DataFrame(columns= ['tweet_id', "username", "created_at",
                                   "hashtags", "text", 
                                   "reply_to_user", "is_retweeted"])

## Collect Tweets
tweets_list = tw.Cursor(api.search, q=search_words, count= 10, lang='en').items()

# Save the copy of an object
tweets_list_copy = tweets_list

### File name to save original tweet
filename = filepath + 'twitter_data_analysis' + (datetime.datetime.now().strftime("%Y-%m-%d-%H")) + '.csv'

for tweet in tweets_list:
    tweet_id = tweet.id
    created_at = tweet.created_at
    username = tweet.user.screen_name
    reply_to_user = tweet.in_reply_to_user_id
    text_encoded = tweet.text.encode('utf-8')
#   text_deecoded = text_encoded.decode('utf-8')
    is_retweeted = tweet.retweeted

    original_hashtags_list = tweet.entities['hashtags']
    trim_hashtags_list = []  
    for hashtag in original_hashtags_list:
        trim_hashtags_list.append(hashtag['text'])
    hashtags = trim_hashtags_list
    print(hashtags)
    
    #Append to dataframe - db_tweets
    data = [tweet_id, username, created_at, hashtags, text_encoded, reply_to_user, is_retweeted]
    
    with open( filename, 'a+', encoding='utf-8', newline='') as csvFile:
        csvWriter = csv.writer(csvFile)
        csvWriter.writerow(data)

db_tweets.head(4)

### Download  Retweet:

### Get the list of tweet Id
import pandas as pd
file_path = "data/Hnin_Tweeter_Api/"
saved_tweet_file = "twitter_data_analysis2021-04-26-13.csv"
downloaded_tweet = pd.read_csv(file_path +saved_tweet_file )

downloaded_tweet.head()

downloaded_tweet.shape

### Filter for original tweets only 
original_tweets_only = downloaded_tweet[downloaded_tweet['reply_to_user'].isnull()]
original_tweets_only.shape

# replying_tweets = downloaded_tweet[downloaded_tweet['reply_to_user'].notnull()]
# replying_tweets.shape

df_tweetid_username = original_tweets_only[['tweet_id', 'username']]
df_tweetid_username.head()

dict_tweet_list = df_tweetid_username.to_dict('record')

dict_tweet_list[4]

filename_reply_tweets =  filepath + 'twitter_data_analysis_reply_tweets' + (datetime.datetime.now().strftime("%Y-%m-%d-%H")) + '.csv'

tweet = dict_tweet_list[4]

tweets_reply_list = tw.Cursor(api.search, q='to:{}'.format(tweet['username']), since_id=tweet['tweet_id'], lang='en').items()

for tweet in dict_tweet_list:
    tweets_reply_list = tw.Cursor(api.search, q='to:{}'.format(tweet['username']), since_id=tweet['tweet_id'], lang='en').items()
       
        
    for reply_tweet in twee ts_reply_list:
        print("reply_tweet", reply_tweet)
        print("==================================")
        print("tweet", tweet)
     
        tweet_id = reply_tweet.id
        tweet_id_string = reply_tweet.id_string
        print("reply_tweet_id", tweet_id)
        reply_to_id = tweet['tweet_id'] #main tweet id (parent)

        created_at = reply_tweet.created_at
        username = reply_tweet.user.screen_name  
        reply_to_user = reply_tweet.in_reply_to_user_id
        print(created_at)     
    
        text_encoded = reply_tweet.text.encode('utf-8')
        print(text_encoded)
        text_decoded = text_encoded.decode('utf-8')
        
        is_retweeted = reply_tweet.retweeted

        original_hashtags_list = reply_tweet.entities['hashtags']
        trim_hashtags_list = []  
        for hashtag in original_hashtags_list:
            trim_hashtags_list.append(hashtag['text'])
        hashtags = trim_hashtags_list
        print(hashtags)

        #Append to dataframe - db_tweets
        data = [tweet_id, reply_to_id, username, created_at, hashtags, text_encoded, reply_to_user, is_retweeted]

        with open( filename_reply_tweets , 'a+', encoding='utf-8', newline='') as csvFile:
            csvWriter = csv.writer(csvFile)
            csvWriter.writerow(data)

for reply_tweet in tweets_reply_list:
        print("reply_tweet", reply_tweet)

## Reference
# https://medium.com/@I_am_milica/the-beginners-guide-to-downloading-twitter-data-using-tweepy-4ec981eaba77
#https://medium.com/@Luca/collecting-old-tweets-with-the-twitter-premium-api-and-python-e52773d094bd
# db_tweets.to_csv(filepath + filename , index = False)
