==> 3_Network-Table-Creation.ipynb <==
Author: Hnin Pwint Tin, 20 May 2021

Social and Information Network Analysis, 2021 Autumn

Master of Data Science and Innovation , University of Technology Sydney

This is the data preparation script part-3 for network data visualisation assessment task - AT3

### Load Original Tweet Data
import pandas as pd

# original_tweet = pd.read_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis2021-05-03-15.csv")

#original_tweet = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis2021-05-11-13.csv")

### Load Reply Dataset

# reply_tweet = pd.read_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis_reply_tweets2021-05-03-16.csv") == Original Data from Twitter
# reply_tweet = pd.read_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis_reply_tweets2021-05-03-16-updated.csv")

reply_tweet = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis_reply_tweets2021-05-11-15-15.csv")

original_tweet.tail()

len(reply_tweet.username.unique())
# reply_tweet.head(2)

# reply_tweet['reply_to_username'] = original_tweet['username'][0]
# reply_tweet.to_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis_reply_tweets2021-05-03-16-updated.csv", index = False)

### Create Network Datasets -  Edge Table Here

#### Create Edge Table , Source are reply-tweet-id , Target are original tweet id

df_edge_table = pd.DataFrame()
df_edge_table['Source'] = reply_tweet['username']
df_edge_table['Target'] = reply_tweet['reply_to_username']
df_edge_table['reply_text'] = reply_tweet['text']

df_edge_table.head()
# df_edge_table.to_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis-05-03-15-edge-table.csv", index=False)
# df_edge_table.to_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis-05-03-15-edge-table-username-based.csv", index=False)

df_edge_table.to_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-edge-table-username-based.csv", index=False)



username = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-edge-table-username-based.csv")
username.head()

username.groupby('Target').count()

### Create Node Table Here

df_node_table = pd.DataFrame()

df_node_table['Id'] = reply_tweet['username']  # Tweet_id
df_node_table['label'] = reply_tweet['text']


df_node_table.head()

main_user = original_tweet[original_tweet['username'] == 'BertyThomas']
df_second = pd.DataFrame()
df_second['Id'] = main_user['username']
df_second['label'] = main_user['text']


df_second.head()
df_node_table = pd.concat([df_node_table, df_second], axis = 0, ignore_index=True)

# df_node_table.to_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis-05-03-15-node-table.csv", index=False)
# df_node_table.to_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis-05-03-15-node-table-username-based.csv", index=False)

df_node_table.to_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-node-table-username-based.csv", index=False)


df_node_table = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-node-table-username-based.csv")

df_node_table.head()
