==> 4_Network-Table-Creation-kaggle-Dataset.ipynb <==
Author: Hnin Pwint Tin, 20 May 2021

Social and Information Network Analysis, 2021 Autumn

Master of Data Science and Innovation , University of Technology Sydney

This is the data preparation script part-4 for network data visualisation assessment task - AT3

import pandas as pd
kaggle_data = pd.read_csv("data/Kaggle/file.csv")
kaggle_data.head()
kaggle_data.shape

filter_data = kaggle_data[kaggle_data['language'] == 'en'] [['conversation_id', 'id', 'created_at', 'user_id', 'username', 'reply_to', 'tweet']]

# filter_data.groupby('conversation_id').count()
len(filter_data['conversation_id'].unique())

import ast

### Convert String {'user_id': '355989081', 'username': 'ANI'} to dictionary to extact user_id and 
filter_data["reply_to_user_id"] = filter_data["reply_to"].map(lambda x: (ast.literal_eval(x))['user_id'])

filter_data["reply_to_username"] = filter_data["reply_to"].map(lambda x: (ast.literal_eval(x))['username'])


filter_data.shape

# Filter Replied to tweets only
bool_series = pd.notnull(filter_data['reply_to_username'])
cleaned_data = filter_data[bool_series]

cleaned_data.shape

# cleaned_data.shape - 69736
len(cleaned_data['conversation_id'].unique()) # 58685

len(cleaned_data['username'].unique())  # 51358

len(cleaned_data['reply_to_username'].unique()) # 40431

### Node Table Creation - usernames

#### Get the list of all usernames to form the node table
userid_list = []
userid_list = cleaned_data['username'].unique()
import numpy as np

userid_list =np.append (userid_list, cleaned_data['reply_to_username'].unique())
print(len(userid_list)) # 91789


df_node_table = pd.DataFrame()
df_node_table['Id'] = userid_list
df_node_table.head()

### Edge Table Creation

df_edge_table = pd.DataFrame()

df_edge_table['Source'] = cleaned_data['username']
df_edge_table['Target'] = cleaned_data['reply_to_username']
df_edge_table['reply_text'] = cleaned_data['tweet']
df_edge_table.head()

jcho_self_edge = df_edge_table[(df_edge_table['Source'] == "jcho710") & (df_edge_table['Target'] == "jcho710")]
print(jcho_self_edge.shape)
jcho_self_edge.head(20)

others_to_jcho_edge = df_edge_table[(df_edge_table['Source'] != "jcho710") & (df_edge_table['Target'] == "jcho710")]
others_to_jcho_edge.shape

### Save the Network Tables

df_edge_table.to_csv("data/Kaggle/twitter_data_analysis-edge-table-username-based.csv", index=False)
df_node_table.to_csv("data/Kaggle/twitter_data_analysis-node-table-username-based.csv", index=False)

df_edge_table.shape

### Unit Testing

import ast

dictionary = ast.literal_eval("{'user_id': '355989081', 'username': 'ANI'}")
print(dictionary)
