==> 5_Emo-Detection.ipynb <==
Author: Hnin Pwint Tin, 20 May 2021

Social and Information Network Analysis, 2021 Autumn

Master of Data Science and Innovation , University of Technology Sydney

This is the data preparation script part-5 for network data visualisation assessment task - AT3

import igraph as ig
import pandas as pd

import matplotlib.pyplot as plt
# Regular Expression Lib
import re

## Load pre-processed data

# edges = pd.read_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis-05-03-15-edge-table-username-based.csv")
# edges = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-edge-table-username-based.csv")
edges = pd.read_csv("data/Kaggle/twitter_data_analysis-edge-table-username-based.csv")


# vertices = pd.read_csv("data/Hnin_Tweeter_Api/may-3/twitter_data_analysis-05-03-15-node-table-username-based.csv")
# vertices = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-node-table-username-based.csv")
vertices = pd.read_csv("data/Kaggle/twitter_data_analysis-node-table-username-based.csv")

edges.head(20)

edges.shape

len(edges['Source'].unique())

len(edges['Target'].unique())

vertices.shape

reply_count_df = edges.groupby(['Target']).count().reset_index()
reply_count_df = reply_count_df.sort_values(by='Source', ascending=False)
reply_count_df.head(20)

g = ig.Graph.DataFrame(edges[['Source', 'Target']], directed=False)

%%time
layout = g.layout(layout='fruchterman_reingold')
layout

fig, ax = plt.subplots(figsize=(20,20))
ig.plot(g, layout=layout, target=ax)

###  Text Data cleaning

#### Removing Http urls
edges["reply_text_cleaned"] = edges['reply_text'].map(lambda x:  re.sub(r'\bhttps?:\/\/.*[\r\n]*', '', x))

# print(edges['reply_text'][4])
print(edges['reply_text_cleaned'][302])

#### Removing mention tag @
edges["reply_text_cleaned"] = edges['reply_text_cleaned'].map(lambda x:  re.sub(r'@[A-Za-z0-9,_]*', '', x))
print(edges['reply_text_cleaned'][302])

#### Removing mention tag #
edges["reply_text_cleaned"] = edges['reply_text_cleaned'].map(lambda x:  re.sub(r'#[A-Za-z0-9,_]*', '', x))
print(edges['reply_text_cleaned'][302])

#### Removing emjii encoding Tried to remove emoji encoding - \xf0\x9f\x99\x8f\xf0\x9f\x99\x8f\xf0\x9f\x99\x8f, but it is wrong approach
edges["reply_text_cleaned"] = edges['reply_text_cleaned'].map(lambda x:  re.sub(r'[\\][x][0-9,a-f,A-F][0-9,a-f,A-F]*', '', x))
print(edges['reply_text_cleaned'][302])
# ([\\][x][0-9,a-f,A-F][0-9,a-f,A-F])

### Decode the text for future emotion text detection for emojii icons
# edges['decoded_text'] = edges['reply_text_cleaned'].map(lambda x: x.encode('utf-8').decode('utf-8'))
# print(edges['decoded_text'][302])

## Exmine the text cleaning result
print(edges['reply_text'][8])
print(edges['reply_text_cleaned'][8])
# print(edges['decoded_text'][8])

#### Store the result here into the Drive

### Decoded cleaned Text to Emotion using text2emotion

import text2emotion as te

## Add Emotion Grade into Dataframe
edges['emotion_grade'] = edges['reply_text_cleaned'].map(lambda x: te.get_emotion(x))

edges.head()

edges.to_csv("data/Kaggle/twitter_data_analysis-edge-table-username-based-emotion.csv", index=False)

edges["emotion_grade_max"] = edges['emotion_grade'].map(lambda x: 'neutral' if ( max(x.values()) == 0)  
                                                        else max(x, key=x.get))

stats = {'key1':20, 'key2':35, 'key3': 44}
max_key = max(stats, key = stats.get)
print(max_key)

#### Drop the mid-processed columns , to save the memory space
edges = edges.drop(columns=['decoded_text'])

# edges.to_csv("data/Hnin_Tweeter_Api/may-11/temp_emotion_result.csv", index= False)
edges.head(10)

### Save the final result 

edges.head(4)
### Save the result
edges.to_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-edge-table-username-based-emotion.csv", index=False)

### Unit Tests

car = {
  "brand": "Ford",
  "model": "Mustang",
  "year": 1964
}

x = car.values()
x

# text = b'@mayareddykp @kpbrindareddy Always welcome sister\xf0\x9f\x99\x8f\xf0\x9f\x99\x8f\xf0\x9f\x99\x8f'
# text = b'@PragativadiNews But online classes at KIIT are still going on. Why?! \xf0\x9f\x98\xb6'
text = b'@Sahana0207Kumar Stay strong India \xf0\x9f\x87\xae\xf0\x9f\x87\xb3'
# text = b' \xf0\x9f\xa5\xb2 Thanks Anna'

decoded_text = text.decode('utf-8')
print(decoded_text)
print(type(decoded_text))
te.get_emotion(decoded_text)

### Copy the emo result to nodes
df_node_table = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-node-table-username-based.csv")



df_node_table.shape
edges.shape

df_node_table["label_cleaned"] = df_node_table['label'].map(lambda x:  re.sub(r'\bhttps?:\/\/.*[\r\n]*', '', x))

df_node_table["label_cleaned"] = df_node_table['label_cleaned'].map(lambda x:  re.sub(r'@[A-Za-z0-9,_]*', '', x))

df_node_table["label_cleaned"] = df_node_table['label_cleaned'].map(lambda x:  re.sub(r'[\\][x][0-9,a-f,A-F][0-9,a-f,A-F]*', '', x))


### Adding "Neutral" emotion when all values are zeros
df_node_table["emotion_grade_max"] = df_node_table['emotion_grade'].map(lambda x: 'neutral' if ( max(x.values()) == 0)  
                                                         else max(x, key=x.get))

### Save the final result to CSV
df_node_table.to_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-node-table-username-based-emotion.csv", index=False)


edge_table = pd.read_csv("data/Hnin_Tweeter_Api/may-11/twitter_data_analysis-05-11-15-edge-table-username-based-emotion.csv")
reply_count_df = edge_table.groupby(['Target']).count().reset_index()
reply_count_df = reply_count_df.sort_values(by='Source', ascending=False)
reply_count_df

edge_table.shape #1723
len(edge_table['Target'].unique()) #15
len(edge_table['Source'].unique()) # 937

### Unit Testing for Emotion

x = {'Happy': 0, 'Angry': 0, 'Surprise': 0, 'Sad': 0, 'Fear': 0}
result =  'neutral' if ( ((max(x, key=x.get) == 0))) else max(x, key=x.get)
result

import text2emotion as te

#test_text_2 = b' \xf0\x9f\xa5\xb2 Thanks Anna'
# test_text_2 = b' Public be like\xf0\x9f\x9a\xb6\xe2\x80\x8d'
# test_text = b' Great work. Your telegram channel is a life saver in this chaos. Thank you for what you have done \xf0\x9f\x99\x8f\xf0\x9f\x8f\xbb'
test_text_2 = b'thank you so much \xf0\x9f\x99\x8f\xf0\x9f\x8f\xbb'

print(type(test_text_2))

decoded_text = test_text_2.decode('utf-8')
print(decoded_text)
print(te.get_emotion(decoded_text))


text = "Day was pretty amazing😃😃"
te.get_emotion(text)

clean_tweets_list = edges['decoded_text']
emotion_list = []

for tweet in clean_tweets_list:
#     decoded_text = test_text.decode('utf-8')
    emotion_list.append(te.get_emotion(tweet))

len(emotion_list)
edges['emotion_result'] = emotion_list

edges['emotion_result'][6]

# emotion_list
print("\U0001F60D")
print("\U0001F601")


### Unit Testing 

a_dictionary = {'Happy': 0.5, 'Angry': 0.5, 'Surprise': 0.0, 'Sad': 0.0, 'Fear': 0.0}

max_key = max(a_dictionary, key=a_dictionary.get)
max_key
